+++ ecommerce/src/main/java/com/example/ecommerce/service/ProductService.java
@@ -8,9 +8,9 @@ public interface ProductService {
 	public Product saveProduct(Product product);
 	public List<Product> saveProducts(List<Product> products);
 	public List<Product> getProducts();
-	public Product getProductById(int id);
+	public Product getProductById(Long id);
 	public Product getProductByName(String name);
-	public String deleteProduct(int id);
+	public String deleteProduct(Long id);
 	public Product updateProduct(Product product);
 }
